move run-task from app.py to audit_lambda.py, for separation
